changed mode::good_strat(s, wd) to mode::good_strat(s) and mode::set_wd(wd); added new profile/results

+++ gen.sh
@@ -6,8 +6,9 @@ M_CODE1=1
 M_CODE2=3
 WORDLEN=10
 
-g++ -std=c++11 src/$EXEC.cpp -o $EXEC
-
-rm -f -r $GFOLDER
-mkdir $GFOLDER
-./$EXEC $GFOLDER/graph $CNT_NODES $CNT_GRAPHS $M_CODE1 $M_CODE2 $WORDLEN
+if g++ -std=c++11 src/$EXEC.cpp -o $EXEC;
+then
+    rm -f -r $GFOLDER
+    mkdir $GFOLDER
+    ./$EXEC $GFOLDER/graph $CNT_NODES $CNT_GRAPHS $M_CODE1 $M_CODE2 $WORDLEN
+fi
